types; refactor
