+++ app/_components/inputs/tests/Governments.test.tsx
@@ -1,6 +1,6 @@
 import { render, screen } from '@testing-library/react';
 import { useForm } from 'react-hook-form';
-import { SystemForm } from '@/app/_types/forms';
+import { SystemForm } from '@/app/_types/system';
 import GovernmentsField from '../Governments';
 import governments from '@/app/_lib/government-list';
