+++ app/_components/inputs/tests/PowerEffects.test.tsx
@@ -1,6 +1,6 @@
 import { render, screen } from '@testing-library/react';
 import { useForm } from 'react-hook-form';
-import { SystemForm } from '@/app/_types/forms';
+import { SystemForm } from '@/app/_types/system';
 import PowerEffectsField from '../PowerEffects';
 
 describe('PowerEffectsField', () => {
