+++ app/_types/forms.ts
@@ -1,18 +1,3 @@
-export type SystemForm = {
-  systemId?: string;
-  onlyPopulated?: boolean;
-  allegiance?: string;
-  government?: string;
-  economy?: string;
-  minorFaction?: string;
-  stationFilter?: string;
-  power?: string;
-  powerEffect?: string;
-  security?: string;
-  factionState?: string;
-  requiresPermit?: boolean;
-};
-
 export type StationForm = {
   stationId?: string;
   ships?: {
