+++ app/_types/system.ts
@@ -1,0 +1,25 @@
+export type SystemForm = {
+  systemId?: string;
+  onlyPopulated?: boolean;
+  allegiance?: string;
+  government?: string;
+  economy?: string;
+  minorFaction?: string;
+  stationFilter?: string;
+  power?: string;
+  powerEffect?: string;
+  security?: string;
+  factionState?: string;
+  requiresPermit?: boolean;
+};
+
+export type System = {
+  name: string;
+  coordinate: {
+    x: number;
+    y: number;
+    z: number;
+  };
+  eliteId: number;
+  starClass: string;
+};
