+++ mock-api/db.json
@@ -15,9 +15,29 @@
   ],
   "comments": [{ "id": 1, "body": "some comment", "postId": 1 }],
   "systems": [
-    { "name": "Sol" },
-    { "name": "Barnards Star" },
-    { "name": "Alpha Centauri" },
-    { "name": "Wolf 359" }
+    {
+      "name": "Sol",
+      "coordinates": { "x": 1, "y": 2, "z": 1 },
+      "eliteId": 111111,
+      "starClass": "k"
+    },
+    {
+      "name": "Barnards Star",
+      "coordinates": { "x": 1, "y": 2, "z": 1 },
+      "eliteId": 111111,
+      "starClass": "k"
+    },
+    {
+      "name": "Alpha Centauri",
+      "coordinates": { "x": 1, "y": 2, "z": 1 },
+      "eliteId": 111111,
+      "starClass": "k"
+    },
+    {
+      "name": "Wolf 359",
+      "coordinates": { "x": 1, "y": 2, "z": 1 },
+      "eliteId": 111111,
+      "starClass": "k"
+    }
   ]
 }
